pyproject.toml
[tool.poetry]
name = "django-music-backend"
version = "0.1.0"
description = ""
authors = ["JackPadalino <pada0867@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
asgiref = "^3.7.2"
django-colorfield = "^0.11.0"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
pytz = "^2023.3.post1"
sqlparse = "^0.4.4"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
